requirements.txt
pystray~=0.16.0
Pillow~=7.2.0